+++ backend/Bartender.Domain/Bartender.Domain.csproj
@@ -7,8 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Bartender.Data\Bartender.Data.csproj" />
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.3.0" />
   </ItemGroup>
 
-    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.3.0" />
+  <ItemGroup>
+    <ProjectReference Include="..\Bartender.Data\Bartender.Data.csproj" />
+  </ItemGroup>
 </Project>
