+++ backend/Bartender.Domain/Interfaces/ICurrentUserContext.cs
@@ -6,6 +6,4 @@ public interface ICurrentUserContext
 {
     int UserId { get; }
     Task<Staff> GetCurrentUserAsync();        // Optionally include Place & Business navigation
-    Task<int> GetUserPlaceIdAsync();
-    Task<int> GetUserBusinessIdAsync();
 }
